requirements.txt
asgiref==3.8.1
Django==4.2.13
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-simple-history==3.7.0
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
